requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-postgres
python-dotenv
pypdf
langchain-core
langchain-postgres
psycopg[binary]